backend/requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Base de données
sqlalchemy==2.0.49
asyncpg==0.31.0
alembic==1.13.1

# Sécurité
python-jose[cryptography]==3.4.0
bcrypt==5.0.0
python-multipart==0.0.26

# Rate Limiting
slowapi==0.1.9

# Config & utilitaires
pydantic-settings==2.2.1
python-dotenv==1.2.2

# Observabilité
loguru==0.7.2
sentry-sdk[fastapi]==2.8.0

# Tests
pytest==9.0.3
pytest-asyncio==0.23.7
pytest-cov==5.0.0
pytest-mock==3.14.0
httpx==0.28.1
aiosqlite==0.22.1

# Qualité & Sécurité
bandit==1.7.9
pip-audit==2.10.0
semgrep==1.72.0
detect-secrets==1.5.0
checkov==3.2.0

# CyberScan SaaS
stripe==10.5.0
apscheduler==3.10.4
reportlab==4.4.10

# Email transactionnel
resend==2.7.0

# 2FA
pyotp==2.9.0
qrcode[pil]==7.4.2

# Dev
pre-commit==4.5.1
